Pul o'tkazmalarini uchun double bigDecimelga o'tkazildi hafsizlik uchun

+++ src/main/java/dev/shoxruhjon/models/Ticket.java
@@ -1,15 +1,16 @@
 package dev.shoxruhjon.models;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class Ticket extends BaseEntity {
     private final String trainId;
     private final String userId;
     private final LocalDateTime purchaseTime;
-    private final double price;
+    private final BigDecimal price;
     private boolean cancelled;
 
-    public Ticket(String trainId, String userId, double price) {
+    public Ticket(String trainId, String userId, BigDecimal price) {
         super();
         this.trainId = trainId;
         this.userId = userId;
@@ -30,7 +31,7 @@ public class Ticket extends BaseEntity {
         return purchaseTime;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
