+++ src/main/java/dev/shoxruhjon/models/WalletTransaction.java
@@ -1,14 +1,15 @@
 package dev.shoxruhjon.models;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class WalletTransaction extends BaseEntity {
     private final String userId;
-    private final double amount;
+    private final BigDecimal amount;
     private final String type;
     private final LocalDateTime timestamp;
 
-    public WalletTransaction(String userId, double amount, String type) {
+    public WalletTransaction(String userId, BigDecimal amount, String type) {
         super();
         this.userId = userId;
         this.amount = amount;
@@ -16,8 +17,19 @@ public class WalletTransaction extends BaseEntity {
         this.timestamp = LocalDateTime.now();
     }
 
-    public String getUserId() { return userId; }
-    public double getAmount() { return amount; }
-    public String getType() { return type; }
-    public LocalDateTime getTimestamp() { return timestamp; }
+    public String getUserId() {
+        return userId;
+    }
+
+    public BigDecimal getAmount() {
+        return amount;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public LocalDateTime getTimestamp() {
+        return timestamp;
+    }
 }
